refactor: improvements from review
